refactor(web/supabase): migrate Supabase client import and enhance client creation
- Updated import paths for the Supabase client to use the new structure in the lib directory.
- Enhanced the createClient function to include error handling for missing environment variables.
- Removed the old Supabase client utility file to clean up the codebase.
